chore: production readiness report chore: tests added chore: smart enum added for PaymentType

+++ Refactoring/Application/Features/ProcessOrder/ProcessOrderCommand.cs
@@ -4,7 +4,7 @@ namespace Refactoring.Application.Features.ProcessOrder;
 
 public record ProcessOrderCommand : IRequest
 {
-    public required int ProductId { get; init; }
+    public required long ProductId { get; init; }
     public required string ProductType { get; init; }
     public required string PaymentType { get; init; }
 }
