+++ Refactoring/Domain/Interfaces/DB/IOrdersRepository.cs
@@ -4,5 +4,5 @@ namespace Refactoring.Domain.Interfaces.DB;
 
 public interface IProductsRepository
 {
-    Task<Product?> GetFirstByIdAndTypeAsync(int id, string type, CancellationToken cancellationToken); //TODO I'd use different ID to communicate with outside world
+    Task<Product?> GetFirstByIdAndTypeAsync(long id, string type, CancellationToken cancellationToken); //TODO I'd use different ID to communicate with outside world
 }
