+++ Refactoring/Infrastructure/DB/ProductsRepository.cs
@@ -6,7 +6,7 @@ namespace Refactoring.Infrastructure.DB;
 
 internal class ProductsRepository(OrdersDbContext dbContext) : IProductsRepository
 {
-    public async Task<Product?> GetFirstByIdAndTypeAsync(int id, string type, CancellationToken cancellationToken)
+    public async Task<Product?> GetFirstByIdAndTypeAsync(long id, string type, CancellationToken cancellationToken)
     {
         return await dbContext.Products.FirstOrDefaultAsync(x => 
             x.Id == id && x.Type == type,
